Rename remaining entities with name "transformer" to "transform"
